# docs/swagger.yaml
definitions:
...
  httputil.Error:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  webhook.Webhook:
    properties:
      createdAt:
        type: string
      enabled:
        type: boolean
      events:
        description: e.g., ["login_success", "register", "token_revoked"]
        items:
          type: string
        type: array
      id:
        type: string
      secret:
        description: HMAC-SHA256 signing secret
        type: string
      tenantID:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /tenants/{tid}/webhooks:
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - AdminAuth: []
      summary: List webhooks
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Register a new webhook endpoint that will receive event notifications
        for the tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Webhook details (url, events[])
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/webhook.Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Create webhook
      tags:
      - webhooks
  /tenants/{tid}/webhooks/{wid}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: tid
        required: true
        type: string
      - description: Webhook ID
        in: path
        name: wid
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.Error'
      security:
      - AdminAuth: []
      summary: Delete webhook
      tags:
      - webhooks
...
swagger: "2.0"